10-20x improvement in determining a producer
So we were iterating through a slice of pointers (one cache miss) to producers
then referencing a pointer to a shard (another cache miss) then generating the
*big.Int key (for each shard) and doing two comparisons (which do allocations)
This is the OGShards below

New strategy uses the fact that we normally have a power of 2 shards, like all of them in us1.
When getting the shards we can know this, then we know we can just shift the correct
number of bits so that we can just index the slice.
NewShards

In the case it's not a power of 2, we're slightly better than OG, we at least don't create
a new key for each partition.
NewShardsNoPower

BenchmarkOGShards-12             	  355597	      4157 ns/op	    1351 B/op	      31 allocs/op
BenchmarkNewAShards-12           	 3972441	       280 ns/op	      99 B/op	       3 allocs/op
BenchmarkNewAShardsNoPower-12    	  352802	      2989 ns/op	    1040 B/op	      26 allocs/op